classifier/requirements.txt
Flask==1.1.1
numpy==1.16.4
pandas==0.24.2
pandas-profiling==2.3.0
Pillow==6.1.0
torch==1.3.0
torchvision==0.4.1
